moved BooleanLiteral
