+++ src_smartcontract/sc/alinouslang.h
@@ -54,7 +54,6 @@
 
 #include "sc_expression/AllocationExpression.h"
 #include "sc_expression/ArrayReferenceExpression.h"
-#include "sc_expression/BooleanLiteral.h"
 #include "sc_expression/CastExpression.h"
 #include "sc_expression/ConstructorArray.h"
 #include "sc_expression/ConstructorCall.h"
@@ -64,6 +63,7 @@
 #include "sc_expression/ParenthesisExpression.h"
 #include "sc_expression/VariableIdentifier.h"
 
+#include "sc_expression_literal/BooleanLiteral.h"
 #include "sc_expression_literal/NullLiteral.h"
 #include "sc_expression_literal/NumberLiteral.h"
