+++ src_smartcontract/sc_expression_literal/CMakeLists.txt
@@ -1,6 +1,7 @@
 
 
 set(__src
+	BooleanLiteral.cpp
 	NullLiteral.cpp
 	NumberLiteral.cpp
 )
